+++ server/src/main/java/com/hedera/block/server/BlockNodeAppInjectionComponent.java
@@ -23,7 +23,6 @@ import com.hedera.block.server.metrics.MetricsInjectionModule;
 import com.hedera.block.server.notifier.NotifierInjectionModule;
 import com.hedera.block.server.persistence.PersistenceInjectionModule;
 import com.hedera.block.server.service.ServiceInjectionModule;
-import com.hedera.block.server.verifier.VerifierInjectionModule;
 import com.swirlds.config.api.Configuration;
 import dagger.BindsInstance;
 import dagger.Component;
@@ -39,7 +38,6 @@ import javax.inject.Singleton;
             HealthInjectionModule.class,
             PersistenceInjectionModule.class,
             MediatorInjectionModule.class,
-            VerifierInjectionModule.class,
             ConfigInjectionModule.class,
             MetricsInjectionModule.class,
         })
