--- server/src/main/java/com/hedera/block/server/verifier/VerifierInjectionModule.java
@@ -1,34 +1,0 @@
-/*
- * Copyright (C) 2024 Hedera Hashgraph, LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.hedera.block.server.verifier;
-
-import com.hedera.block.server.events.BlockNodeEventHandler;
-import com.hedera.block.server.events.ObjectEvent;
-import com.hedera.hapi.block.SubscribeStreamResponse;
-import dagger.Binds;
-import dagger.Module;
-import javax.inject.Singleton;
-
-/** A Dagger module for providing dependencies for Validator Module. */
-@Module
-public interface VerifierInjectionModule {
-
-    @Binds
-    @Singleton
-    BlockNodeEventHandler<ObjectEvent<SubscribeStreamResponse>> bindBlockNodeEventHandler(
-            StreamPersistenceHandlerImpl streamPersistenceHandler);
-}
