+++ server/src/main/java/module-info.java
@@ -15,7 +15,6 @@ module com.hedera.block.server {
     exports com.hedera.block.server.events;
     exports com.hedera.block.server.health;
     exports com.hedera.block.server.persistence;
-    exports com.hedera.block.server.verifier;
     exports com.hedera.block.server.notifier;
     exports com.hedera.block.server.service;
