+++ server/src/test/java/com/hedera/block/server/persistence/StreamPersistenceHandlerImplTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hedera.block.server.verifier;
+package com.hedera.block.server.persistence;
 
 import static com.hedera.block.server.metrics.BlockNodeMetricTypes.Counter.LiveBlocksVerified;
 import static com.hedera.block.server.util.PersistTestUtils.generateBlockItems;
